generateSample uses component content
Make the `generateSample` function a bit more dynamic in concept, by
allowing it to take as input the component (and a marker). The function
will render the component’s content to string, then truncate it, and
use that as a “sample” that is returned. Doing this allows the
framework to generate a sample rather than it being each post’s
responsibility.